Try harder to ensure that the child process dies.
This is done by trapping and propagating SIGINT/SIGTERM where possible,
propagating them to the child, and then following up with a SIGKILL
after a delay.

I have no idea what happens on Windows since sending SIGINT isn't
possible per the Go documentation, but I assume that ProcessTerminate()
will get called after the delay elapses.

Currently the delay is set to 5 seconds.  This could be longer and or
configurable but the only thing I know of that needs a longer wait is
Tor, and people should be killing the child instead of the wrapper
anyway.